requirements.txt
pyupbit==0.2.31
openai==1.3.0
python-dotenv==1.0.0
requests==2.31.0
pandas==2.1.3
numpy==1.25.2